+++ R/list.R
@@ -18,6 +18,8 @@
 #'   the filenames.
 #' @inheritParams path_filter
 #' @param all If `TRUE` hidden files are also returned.
+#' @param fail Should the call fail (the default) or warn if a file cannot be
+#'   accessed.
 #' @template fs
 #' @export
 #' @examples
@@ -39,14 +41,14 @@
 #' link_delete("base")
 #' \dontshow{setwd(.old_wd)}
 dir_ls <- function(path = ".", all = FALSE, recursive = FALSE, type = "any",
-                   glob = NULL, regexp = NULL, invert = FALSE, ...) {
+                   glob = NULL, regexp = NULL, invert = FALSE, fail = TRUE, ...) {
   assert_no_missing(path)
 
   old <- path_expand(path)
 
-  files <- as.character(dir_map(old, identity, all, recursive, type))
+  files <- as.character(dir_map(old, identity, all, recursive, type, fail))
 
-  path_filter(files, glob, regexp, invert = invert,...)
+  path_filter(files, glob, regexp, invert = invert, ...)
 }
 
 directory_entry_types <- c(
@@ -63,33 +65,33 @@ directory_entry_types <- c(
 #' @rdname dir_ls
 #' @param fun A function, taking one parameter, the current path entry.
 #' @export
-dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
+dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any", fail = TRUE) {
   assert_no_missing(path)
 
   type <- match.arg(type, names(directory_entry_types), several.ok = TRUE)
 
   old <- path_expand(path)
 
-  dir_map_(old, fun, all, sum(directory_entry_types[type]), recursive)
+  dir_map_(old, fun, all, sum(directory_entry_types[type]), recursive, fail)
 }
 
 #' @rdname dir_ls
 #' @export
-dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
+dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any", fail = TRUE) {
   assert_no_missing(path)
 
   old <- path_expand(path)
 
-  dir_map(old, fun, all, recursive, type)
+  dir_map(old, fun, all, recursive, type, fail)
   invisible(path_tidy(path))
 }
 
 #' @rdname dir_ls
 #' @export
 dir_info <- function(path = ".", all = FALSE, recursive = FALSE,
-                     type = "any", regexp = NULL, glob = NULL, ...) {
+                     type = "any", regexp = NULL, glob = NULL, fail = TRUE, ...) {
   assert_no_missing(path)
 
   file_info(dir_ls(path = path, all = all, recursive = recursive, type = type,
-    regexp = regexp, glob = glob, ...))
+    regexp = regexp, glob = glob, fail = fail, ...), fail = fail)
 }