+++ src/error.cc
@@ -2,12 +2,13 @@
 
 #define BUFSIZE 8192
 
-SEXP error_condition(uv_fs_t req, const char* loc, const char* format, ...) {
+bool signal_condition(
+    uv_fs_t req, const char* loc, bool error, const char* format, ...) {
   SEXP condition, c, signalConditionFun, out;
   va_list ap;
 
   if (req.result >= 0) {
-    return R_NilValue;
+    return false;
   }
   int err = req.result;
   uv_fs_req_cleanup(&req);
@@ -18,7 +19,7 @@ SEXP error_condition(uv_fs_t req, const char* loc, const char* format, ...) {
   PROTECT(c = Rf_allocVector(STRSXP, 4));
   SET_STRING_ELT(c, 0, Rf_mkChar(uv_err_name(err)));
   SET_STRING_ELT(c, 1, Rf_mkChar("fs_error"));
-  SET_STRING_ELT(c, 2, Rf_mkChar("error"));
+  SET_STRING_ELT(c, 2, Rf_mkChar(error ? "error" : "warning"));
   SET_STRING_ELT(c, 3, Rf_mkChar("condition"));
 
   char buf[BUFSIZE];
@@ -32,12 +33,13 @@ SEXP error_condition(uv_fs_t req, const char* loc, const char* format, ...) {
   SET_VECTOR_ELT(condition, 0, Rf_mkString(buf));
   Rf_setAttrib(condition, R_ClassSymbol, c);
   Rf_setAttrib(condition, Rf_mkString("location"), Rf_mkString(loc));
-  signalConditionFun = Rf_findFun(Rf_install("stop"), R_BaseEnv);
+  signalConditionFun =
+      Rf_findFun(Rf_install(error ? "stop" : "warning"), R_BaseEnv);
 
   SEXP call = PROTECT(Rf_lang2(signalConditionFun, condition));
   PROTECT(out = Rf_eval(call, R_GlobalEnv));
 
   UNPROTECT(4);
 
-  return out;
+  return true;
 }