+++ tests/testthat/test-list.R
@@ -72,6 +72,19 @@ describe("dir_ls", {
   it("errors on missing input", {
     expect_error(dir_ls(NA), class = "invalid_argument")
   })
+  it("warns if fail == FALSE", {
+    with_dir_tree(list(
+        "foo",
+        "foo2/bar/baz"), {
+      file_chmod("foo", "a-r")
+
+      expect_error(dir_ls(".", recursive = TRUE), class = "EACCES")
+      expect_warning(dir_ls(fail = FALSE, recursive = TRUE), class = "EACCES")
+
+      file_chmod("foo2/bar", "a-r")
+      expect_warning(dir_ls("foo2", fail = FALSE, recursive = TRUE), class = "EACCES")
+    })
+  })
 })
 
 describe("dir_map", {
@@ -87,6 +100,20 @@ describe("dir_map", {
   })
   it("errors on missing input", {
     expect_error(dir_map(NA, fun = identity), class = "invalid_argument")
+  })
+
+  it("warns if fail == FALSE", {
+    with_dir_tree(list(
+        "foo",
+        "foo2/bar/baz"), {
+      file_chmod("foo", "a-r")
+
+      expect_error(dir_map(".", fun = identity, recursive = TRUE), class = "EACCES")
+      expect_warning(dir_map(fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+
+      file_chmod("foo2/bar", "a-r")
+      expect_warning(dir_map("foo2", fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+    })
   })
 })
 
@@ -121,6 +148,20 @@ describe("dir_walk", {
   it("errors on missing input", {
     expect_error(dir_walk(NA, fun = identity), class = "invalid_argument")
   })
+
+  it("warns if fail == FALSE", {
+    with_dir_tree(list(
+        "foo",
+        "foo2/bar/baz"), {
+      file_chmod("foo", "a-r")
+
+      expect_error(dir_walk(".", fun = identity, recursive = TRUE), class = "EACCES")
+      expect_warning(dir_walk(fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+
+      file_chmod("foo2/bar", "a-r")
+      expect_warning(dir_walk("foo2", fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+    })
+  })
 })
 
 describe("dir_info", {
@@ -134,5 +175,19 @@ describe("dir_info", {
   })
   it("errors on missing input", {
     expect_error(dir_info(NA), class = "invalid_argument")
+  })
+
+  it("warns if fail == FALSE", {
+    with_dir_tree(list(
+        "foo",
+        "foo2/bar/baz"), {
+      file_chmod("foo", "a-r")
+
+      expect_error(dir_info(".", fun = identity, recursive = TRUE), class = "EACCES")
+      expect_warning(dir_info(fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+
+      file_chmod("foo2/bar", "a-r")
+      expect_warning(dir_info("foo2", fail = FALSE, fun = identity, recursive = TRUE), class = "EACCES")
+    })
   })
 })